+++ tests/testthat/test-random.R
@@ -91,3 +91,30 @@ test_with_dir("Random targets are reproducible", {
   expect_false(identical(readd(my), old_my))
   expect_false(identical(readd(mz), old_mz))
 })
+
+test_that("random seed can be read", {
+  cache <- storr::storr_environment() # For the examples.
+  my_plan <- drake_plan(
+    target1 = sqrt(1234),
+    target2 = rnorm(n = 1, mean = target1)
+  )
+  digest::digest(.Random.seed) # Take the fingerprint of the current seed.
+  make(my_plan, cache = cache, session_info = FALSE)
+  rs0 <- digest::digest(.Random.seed) # make() protects your R session's seed from being changed.
+  ds0 <- digest::digest(read_drake_seed(cache = cache))
+  targ0 <- readd(target2, cache = cache) # Here is some randomly-generated data.
+  clean(target2, cache = cache) # Oops, I removed the data.
+  x <- runif(1) # Maybe I also changed the R session's seed.
+  rs1 <- digest::digest(.Random.seed) # Different from before.
+  make(my_plan, cache = cache, session_info = FALSE)
+  ds2 <- digest::digest(read_drake_seed(cache = cache))
+  rs2 <- digest::digest(.Random.seed)
+  targ2 <- readd(target2, cache = cache)
+  
+  expect_equal(rs1, rs2)
+  expect_false(rs0 == rs1)
+  expect_equal(ds0, ds2)
+  expect_equal(targ0, targ2)
+  cache$del(key = "seed", namespace = "config")
+  expect_error(read_drake_seed(cache = cache), regexp = "seed not found")
+})